Damn you ~~Jenkins~~ GH Actions!
