+++ engine/test/dep-sorting.ts
@@ -1,4 +1,4 @@
-import { sortModules, } from '../src/router/module'
+import { sortModules, } from '../src/services/mod-sort'
 import { ModuleInterface, } from '../src/modules/interfaces'
 
 const fakeMod = (name: string, dependencies: string[]) => ({
